requirements.txt
django==1.9.7
psycopg2
djangorestframework==3.4.7
markdown
django-filter